P1 correctness fixes: surface Firestore/downstream errors, chunk batches, guard NPEs
Financial read paths no longer mask failures as empty data (an outage would
otherwise render as "customer owes 0"), Firestore batches are chunked to the
500-op limit, and several latent NPE/overflow/aggregation-abort bugs are fixed.

- Repositories: read paths feeding financial totals (PaymentRepository,
  CustomerDebtSummaryRepository.findAll, ManualCashPaymentRepository) now throw
  on Firestore failure instead of returning empty. getAllUniqueCodes* throwing
  also prevents silent duplicate re-imports on a transient read error.
- CustomerAnalysisService / AuditControlService: aggregate downstream fetches
  raise ExternalServiceException on outage; per-customer / no-data paths are
  left tolerant (a 404 there is legitimate).
- PaymentRepository.saveAll & CustomerDebtSummaryRepository.saveAll: chunk to
  Firestore's 500-operation WriteBatch limit.
- PaymentStatusService: include manual-cash payments so cash-only customers are
  not wrongly flagged red.
- AggregationService: guard malformed initial-debt date so one bad row cannot
  abort the whole aggregation job.
- ManualCashPaymentRepository: guard null amount to avoid NPE.
- UniqueCodeGenerator / PaymentRepository / DeduplicationService: amount and
  balance cents use long to avoid int overflow on large values.
- PaymentController: log the previously-swallowed post-dedup aggregation failure.

Adds the architecture audit report under .architecture-audit/.
Full reactor `mvn clean verify` passes (all modules, all tests).

+++ common/src/main/java/ge/tastyerp/common/util/UniqueCodeGenerator.java
@@ -73,7 +73,7 @@ public final class UniqueCodeGenerator {
         }
         return String.valueOf(amount.multiply(BigDecimal.valueOf(100))
                 .setScale(0, java.math.RoundingMode.HALF_UP)
-                .intValue());
+                .longValue());
     }
 
     /**
